+++ src/Types/Guest.tsx
@@ -1,0 +1,5 @@
+export interface Guest {
+  id: number;
+  full_name: string;
+  guest_id: string;
+}
