+++ src/Types/TimeSetting.tsx
@@ -1,0 +1,5 @@
+export interface TimeSetting {
+  id: number;
+  time_amount: number,
+  increment: number,
+}
