+++ src/modules/App/AppProvider.tsx
@@ -2,6 +2,7 @@ import { PropsWithChildren, createContext, useState, useEffect } from "react";
 import { GetLoggedInUser } from "./actions/get-current-user";
 import { getMyChessClubs } from "../../ServerManager";
 import { AppStateDefaults } from "./state";
+import type { ChessClub } from "../../Types/ChessClub";
 
 export const AppContext = createContext(null as any);
 
@@ -9,7 +10,7 @@ interface AppProviderProps extends PropsWithChildren<{}> { };
 
 export function AppProvider({ children }: AppProviderProps) {
   const localVillagerUser = GetLoggedInUser();
-  const [myChessClubs, setMyChessClubs] = useState([]);
+  const [myChessClubs, setMyChessClubs] = useState<ChessClub[]>([]);
   useEffect(
     () => {
       getMyChessClubs()
