+++ src/modules/ChessClubs/actions/handle-form-change.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent } from "react";
 // import type { ChessClub } from "../../App/types";
-import type { ChessClub } from "../types/ChessClub";
+import type { ChessClub } from "../../../Types/ChessClub";
 interface handlerInput {
   stateObject: Partial<ChessClub>;
   handler: React.Dispatch<React.SetStateAction<any>>;
