+++ src/modules/ChessClubs/components/CreateClub.tsx
@@ -1,6 +1,6 @@
 import { createNewClub } from "../../../ServerManager";
 import { Dispatch, SetStateAction, useRef } from "react";
-import type { ChessClubCreate } from "../types/ChessClub";
+import type { ChessClubCreate } from "../../../Types/ChessClub";
 import { handleFormChange } from "../actions/handle-form-change";
 
 interface CreateClubFormProps {
