+++ src/modules/Tournament/Tournament.tsx
@@ -1,17 +1,20 @@
 import { getClubMatesAndGuests, getMyOpenTournaments, getMyPastTournaments } from "../../ServerManager";
 import { useEffect, useState, useContext } from "react";
 import { AppContext } from "../App/AppProvider";
-import type { Player, Guest, NewTournament, Tournament, TimeSetting } from "./Types";
+import type { Guest } from "../../Types/Guest";
 import type { Match } from "tournament-pairings/dist/Match";
 import { ActiveTournament } from "./components/ActiveTournament";
 import { TournamentForm } from "./components/TournamentForm";
 import { TournamentList } from "./components/TournamentList";
 import "../../styles/Tournament.css";
-import { ChessClub } from "../App/types";
+import type { ChessClub } from "../../Types/ChessClub";
+import type { PlayerRelated } from "../../Types/Player";
+import type { TimeSetting } from "../../Types/TimeSetting";
+import type { Tournament, NewTournament } from "../../Types/Tournament";
 
 export const TournamentController = () => {
   const { localVillagerUser, myChessClubs } = useContext(AppContext);
-  const [clubMatesAndGuests, setClubMatesAndGuests] = useState<(Player | Guest)[]>([]);
+  const [clubMatesAndGuests, setClubMatesAndGuests] = useState<(PlayerRelated | Guest)[]>([]);
   const [timeSettings, setTimeSettings] = useState<TimeSetting[]>([]);
   const [myOpenTournaments, setMyOpenTournaments] = useState<Tournament[]>([]);
   const [createNewTournament, setCreateNewTournament] = useState(false);
