+++ src/utils/player-guest-typing.tsx
@@ -1,6 +1,7 @@
-import type { Player, PlayerOnTournament, Guest } from "../modules/Tournament/Types";
+import type { Player, PlayerRelated } from "../Types/Player";
+import type { Guest } from "../Types/Guest";
 
-export const getPlayerType = (player: Player | PlayerOnTournament | Guest) => {
+export const getPlayerType = (player: Player | PlayerRelated | Guest) => {
   if ('guest_id' in player) {
     return 'Guest';
   } else if ('my_clubs' in player) {
